test(adapters-opencode): split adapter index suite
